move the related subtensor tests
